refactor(server): guards, decorators, and utils
