+++ server/src/domain/album/dto/get-albums.dto.ts
@@ -1,8 +1,7 @@
-import { ValidateUUID } from '@app/immich/decorators/validate-uuid.decorator';
-import { toBoolean } from '@app/immich/utils/transform.util';
 import { ApiProperty } from '@nestjs/swagger';
 import { Transform } from 'class-transformer';
 import { IsBoolean, IsOptional } from 'class-validator';
+import { toBoolean, ValidateUUID } from '../../domain.util';
 
 export class GetAlbumsDto {
   @IsOptional()
