+++ server/src/domain/asset/dto/download.dto.ts
@@ -1,6 +1,6 @@
-import { ValidateUUID } from '@app/immich/decorators/validate-uuid.decorator';
 import { ApiProperty } from '@nestjs/swagger';
 import { IsInt, IsOptional, IsPositive } from 'class-validator';
+import { ValidateUUID } from '../../domain.util';
 
 export class DownloadDto {
   @ValidateUUID({ each: true, optional: true })
