+++ server/src/domain/asset/dto/map-marker.dto.ts
@@ -1,7 +1,7 @@
-import { toBoolean } from '@app/immich/utils/transform.util';
 import { ApiProperty } from '@nestjs/swagger';
 import { Transform, Type } from 'class-transformer';
 import { IsBoolean, IsDate, IsOptional } from 'class-validator';
+import { toBoolean } from '../../domain.util';
 
 export class MapMarkerDto {
   @ApiProperty()
