+++ server/src/domain/user/dto/update-user.dto.ts
@@ -1,7 +1,7 @@
-import { toEmail, toSanitized } from '@app/immich/utils/transform.util';
 import { ApiProperty } from '@nestjs/swagger';
 import { Transform } from 'class-transformer';
 import { IsBoolean, IsEmail, IsNotEmpty, IsOptional, IsString, IsUUID } from 'class-validator';
+import { toEmail, toSanitized } from '../../domain.util';
 
 export class UpdateUserDto {
   @IsOptional()
