+++ server/src/immich/api-v1/album/dto/add-users.dto.ts
@@ -1,4 +1,4 @@
-import { ValidateUUID } from '@app/immich/decorators/validate-uuid.decorator';
+import { ValidateUUID } from '@app/domain';
 
 export class AddUsersDto {
   @ValidateUUID({ each: true })
