--- server/src/immich/decorators/use-validation.decorator.ts
@@ -1,12 +1,0 @@
-import { applyDecorators, UsePipes, ValidationPipe } from '@nestjs/common';
-
-export function UseValidation() {
-  return applyDecorators(
-    UsePipes(
-      new ValidationPipe({
-        transform: true,
-        whitelist: true,
-      }),
-    ),
-  );
-}
