--- server/src/immich/utils/path-form-data.util.ts
@@ -1,3 +1,0 @@
-export function patchFormData(latin1: string) {
-  return Buffer.from(latin1, 'latin1').toString('utf8');
-}
