--- server/src/immich/utils/transform.util.ts
@@ -1,18 +1,0 @@
-import sanitize from 'sanitize-filename';
-
-interface IValue {
-  value?: string;
-}
-
-export const toBoolean = ({ value }: IValue) => {
-  if (value == 'true') {
-    return true;
-  } else if (value == 'false') {
-    return false;
-  }
-  return value;
-};
-
-export const toEmail = ({ value }: IValue) => value?.toLowerCase();
-
-export const toSanitized = ({ value }: IValue) => sanitize((value || '').replace(/\./g, ''));
